custom message class extending abstract message class
